Create script to modify cfg-files via ENV
at startup depending on the role this script changes the daemon's cfg files to reflect the environment. This is usually passed on the container via ROLE set in docker-compose.yml.
For starting e.g. multiple chunkserver daemons with different labels on the same host, override LABEL by passing it as command line parameter